requirements.txt
streamlit
pandas
python-dotenv

langchain
langchain-core
langchain-experimental
langchain-google-genai

google-generativeai
tabulate



